Refactors contract concerns out to validation methods

+++ lib/process_settings/replace_versioned_file.rb
@@ -15,11 +15,11 @@ module ProcessSettings
       #   source_file_path must exist on filesystem
       #   source file version has to greater than or equal to destination version
       def replace_file_on_newer_file_version(source_file_path, destination_file_path)
-        source_file_path.to_s      == '' and raise ArgumentError, "source_file_path not present"
-        destination_file_path.to_s == '' and raise ArgumentError, "destination_file_path not present"
-        File.exist?(source_file_path) or raise FileDoesNotExistError, "source file '#{source_file_path}' does not exist"
+        validate_arguments_are_passed_in(source_file_path, destination_file_path)
+        validate_source_file_exists_on_system(source_file_path)
+        validate_source_version_is_not_older(source_file_path, destination_file_path)
 
-        if !File.exist?(destination_file_path) || source_version_is_newer?(source_file_path, destination_file_path)
+        if source_version_is_newer?(source_file_path, destination_file_path)
           FileUtils.mv(source_file_path, destination_file_path)
         elsif source_file_path != destination_file_path # make sure we're not deleting destination file
           FileUtils.rm_f(source_file_path) # clean up, remove left over file
@@ -29,17 +29,37 @@ module ProcessSettings
       private
 
       def source_version_is_newer?(source_file_path, destination_file_path)
-        source_version      = ProcessSettings::TargetedSettings.from_file(source_file_path, only_meta: true).version
-        destination_version = ProcessSettings::TargetedSettings.from_file(destination_file_path, only_meta: true).version
+        if File.exist?(destination_file_path)
+          source_version      = ProcessSettings::TargetedSettings.from_file(source_file_path, only_meta: true).version
+          destination_version = ProcessSettings::TargetedSettings.from_file(destination_file_path, only_meta: true).version
 
-        if Gem::Version.new(source_version) > Gem::Version.new(destination_version)
+          Gem::Version.new(source_version) > Gem::Version.new(destination_version)
+        else
           true
-        elsif Gem::Version.new(source_version) < Gem::Version.new(destination_version)
-          FileUtils.rm_f(source_file_path) # clean up, remove left over file
+        end
+      end
 
-          raise SourceVersionOlderError,
-                "source file '#{source_file_path}' is version #{source_version}"\
-                " and destination file '#{destination_file_path}' is version #{destination_version}"
+      def validate_source_file_exists_on_system(source_file_path)
+        File.exist?(source_file_path) or raise FileDoesNotExistError, "source file '#{source_file_path}' does not exist"
+      end
+
+      def validate_arguments_are_passed_in(source_file_path, destination_file_path)
+        source_file_path.to_s      == '' and raise ArgumentError, "source_file_path not present"
+        destination_file_path.to_s == '' and raise ArgumentError, "destination_file_path not present"
+      end
+
+      def validate_source_version_is_not_older(source_file_path, destination_file_path)
+        if File.exist?(destination_file_path)
+          source_version      = ProcessSettings::TargetedSettings.from_file(source_file_path, only_meta: true).version
+          destination_version = ProcessSettings::TargetedSettings.from_file(destination_file_path, only_meta: true).version
+
+          if Gem::Version.new(source_version) < Gem::Version.new(destination_version)
+            FileUtils.rm_f(source_file_path) # clean up, remove left over file
+
+            raise SourceVersionOlderError,
+                  "source file '#{source_file_path}' is version #{source_version}"\
+                  " and destination file '#{destination_file_path}' is version #{destination_version}"
+          end
         end
       end
     end
